+++ src/ui/monitor/mainWindow.h
@@ -38,7 +38,8 @@
 
 
 class MainWindow : public QMainWindow {
- Q_OBJECT   
+ Q_OBJECT
+
  public:
     MainWindow();
     void informWarningPresence();
